{{cookiecutter.project_slug}}/requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==12.1.0 # pyup: != 11.2.0  # https://github.com/python-pillow/Pillow
{%- if cookiecutter.frontend_pipeline == 'Django Compressor' %}
{%- if cookiecutter.windows == 'y' and cookiecutter.use_docker == 'n' %}
rcssmin==1.2.2 --install-option="--without-c-extensions"  # https://github.com/ndparker/rcssmin
{%- else %}
rcssmin==1.2.2  # https://github.com/ndparker/rcssmin
{%- endif %}
{%- endif %}
argon2-cffi==25.1.0  # https://github.com/hynek/argon2_cffi
{%- if cookiecutter.use_whitenoise == 'y' %}
whitenoise==6.11.0  # https://github.com/evansd/whitenoise
{%- endif %}
redis==7.1.0  # https://github.com/redis/redis-py
{%- if cookiecutter.use_docker == "y" or cookiecutter.windows == "n" %}
hiredis==3.3.0  # https://github.com/redis/hiredis-py
{%- endif %}
{%- if cookiecutter.use_celery == "y" %}
celery==5.6.2  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.8.1  # https://github.com/celery/django-celery-beat
{%- if cookiecutter.use_docker == 'y' %}
flower==2.0.1  # https://github.com/mher/flower
{%- endif %}
{%- endif %}
{%- if cookiecutter.use_async == 'y' %}
uvicorn[standard]==0.40.0  # https://github.com/Kludex/uvicorn
uvicorn-worker==0.4.0  # https://github.com/Kludex/uvicorn-worker
{%- endif %}

# Django
# ------------------------------------------------------------------------------
django==5.2.10  # pyup: < 6.0 # https://www.djangoproject.com/
django-environ==0.12.0  # https://github.com/joke2k/django-environ
django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==65.13.1  # https://github.com/pennersr/django-allauth
{%- if cookiecutter.use_crispy_with_bootstrap == 'y' %}
django-crispy-forms==2.5  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2025.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
{%- endif %}
django-tables2==2.8.0
django-htmx==1.27.0
django-simple-history==3.11.0
#django-cotton
#django-stronghold #todo: replace! https://stackoverflow.com/questions/48407790/restricting-all-the-views-to-authenticated-users-in-django
django-guid==3.5.2
#django-template-partials
django-widget-tweaks==1.5.0 #not updated?
django-tailwind[cookiecutter,reload]==4.4.2
django-hosts==7.0.0 #https://pypi.org/project/django-hosts/ #https://dev.to/ksharma20/how-to-use-subdomains-for-different-apps-in-django-project--fbi
#django-debug-toolbar # already in cookiecutter?
pyjwt[crypto]>=2.8
{%- if cookiecutter.frontend_pipeline == 'Django Compressor' %}
django-compressor==4.6.0  # https://github.com/django-compressor/django-compressor
{%- endif %}
django-redis==6.0.0  # https://github.com/jazzband/django-redis
{%- if cookiecutter.use_drf == 'y' %}
# Django REST Framework
djangorestframework==3.16.1  # https://github.com/encode/django-rest-framework
django-cors-headers==4.9.0  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.29.0  # https://github.com/tfranzel/drf-spectacular
{%- endif %}
{%- if cookiecutter.frontend_pipeline == 'Webpack' %}
django-webpack-loader==3.2.3  # https://github.com/django-webpack/django-webpack-loader
{%- endif %}
# custom:
stripe==14.1.0
